refactor(capture): extract the privacy gate to a shared module
Move the fail-closed privacy gate (the body scan and the metadata token loader)
out of the open step into a shared module so the open step and the upcoming
harvest-time scan use one source of truth and cannot drift. Behavior is
unchanged: the open step still scans the authored body and still fails closed
when the private-name set cannot be loaded.
